feat(core): Add app routing for auth and sessions
