+++ package.json
@@ -21,6 +21,8 @@
   },
   "dependencies": {
     "@expo-google-fonts/space-grotesk": "^0.4.1",
+    "@react-navigation/native": "^7.1.31",
+    "@react-navigation/native-stack": "^7.14.2",
     "expo": "~54.0.0",
     "expo-crypto": "~15.0.8",
     "expo-font": "^55.0.4",
@@ -32,6 +34,8 @@
     "react-dom": "19.1.0",
     "react-native": "0.81.4",
     "react-native-css-interop": "0.2.2",
+    "react-native-safe-area-context": "~5.6.2",
+    "react-native-screens": "~4.16.0",
     "react-native-web": "^0.21.2",
     "tailwindcss": "^3.4.19",
     "tweetnacl": "^1.0.3"
