+++ src/router/AppRouter.tsx
@@ -1,0 +1,25 @@
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+import { LoginScreen } from "../screens/auth/LoginScreen";
+import { SessionsScreen } from "../screens/internal/SessionsScreen";
+import type { RootStackParamList } from "./types";
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+export function AppRouter() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        initialRouteName="auth/login"
+        screenOptions={{
+          headerShown: false,
+          contentStyle: { backgroundColor: "#05050A" },
+        }}
+      >
+        <Stack.Screen name="auth/login" component={LoginScreen} />
+        <Stack.Screen name="internal/sessions" component={SessionsScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
