+++ src/router/types.ts
@@ -1,0 +1,4 @@
+export type RootStackParamList = {
+  "auth/login": undefined;
+  "internal/sessions": undefined;
+};
